+++ crates/nu-command/src/strings/format/mod.rs
@@ -1,0 +1,3 @@
+mod date;
+
+pub use date::FormatDate;
