+++ crates/nu-command/src/strings/mod.rs
@@ -1,6 +1,7 @@
 mod char_;
 mod detect_columns;
 mod encode_decode;
+mod format;
 mod parse;
 mod size;
 mod split;
@@ -9,6 +10,7 @@ mod str_;
 pub use char_::Char;
 pub use detect_columns::*;
 pub use encode_decode::*;
+pub use format::FormatDate;
 pub use parse::*;
 pub use size::Size;
 pub use split::*;
