+++ crates/nu-std/std/log.nu
@@ -139,7 +139,7 @@ def current-log-level [] {
 }
 
 def now [] {
-    date now | date format "%Y-%m-%dT%H:%M:%S%.3f"
+    date now | format date "%Y-%m-%dT%H:%M:%S%.3f"
 }
 
 def handle-log [
