+++ src/tests/test_type_check.rs
@@ -32,14 +32,14 @@ fn number_float() -> TestResult {
 
 #[test]
 fn date_minus_duration() -> TestResult {
-    let input = "2023-04-22 - 2day | date format %Y-%m-%d";
+    let input = "2023-04-22 - 2day | format date %Y-%m-%d";
     let expected = "2023-04-20";
     run_test(input, expected)
 }
 
 #[test]
 fn date_plus_duration() -> TestResult {
-    let input = "2023-04-18 + 2day | date format %Y-%m-%d";
+    let input = "2023-04-18 + 2day | format date %Y-%m-%d";
     let expected = "2023-04-20";
     run_test(input, expected)
 }
